Yet another perl5 solutions for day 22
Annoyingly perl6 was being slow, and I really liked the way it turned
out.  I think there are still bugs in it because I didn't quite read the
docs properly.  I had to grab another implementation off the internet
and figure out what the correct answer was supposed to be.