Day 6, 2018
This day was super annoying as, although this gets the "correct" answer,
I think just seeing if we hit the boundary of our square is an
innacurate way to decide if an area is infinite.   But, whatev, it's
what other folks did and seems to work well enough to let me through.