Day 7, 2018
I at first thought this was just "tsort" plus alphanumeric, but there
was the magic that as soon as any new items came available, you had to
pick the next sorted one.   So if you had AEF available, and complete A,
which frees up B, you then need to work on B, not E.  Sheesh.

The second part was easier tha way though, I guess.