Day 10, 2018
Super annoying, still didn't find a really good way to determine whether
it was the correct second.   Started by just printing all of them that
seemed close (fit in a small bounding box), but ended up looking for the
smallest bounding box (which failed) and then for the smallest "y
range", which worked.   But, nothing says it had to.